Round robin (#16)
* experiment : push new event for exchange subs-true-count times

* fix : round-robin behavior needs to set objectId for every sent message, because we can get next message from common cache which was created with another consumer
new : add ISO-formatted string for decaf::util::Date

* revert : postgresql

* refactoring and renaming

* fix : build, macro __PRETTY_FUNCTION__ should be std::string otherwise Poco::Any can't convert

* fix : add definition POCO_VERSION_PATCH
fix : use different syntax for Poco::Data::SQLite::Utility::throwException

* try to add parallel tests

* revert parallel test

* try to separate every test

+++ libs/libupmq/transport/SimplePriorityMessageDispatchChannel.cpp
@@ -69,7 +69,7 @@ Pointer<Command> SimplePriorityMessageDispatchChannel::dequeue(long long timeout
       if (timeout == -1) {
         mutex.wait();
       } else {
-        mutex.wait(static_cast<unsigned long>(timeout));
+        mutex.wait(timeout);
         break;
       }
     }
